+++ scripts/install-dependencies.sh
@@ -1,9 +1,14 @@
 #!/bin/bash
 
+if [ $# -eq 0 ]
+  then
+    echo "No arguments supplied. Please supply your reachable network interface e.g. eth0 or eth1"
+    exit 1
+fi
+
 pwd=$(pwd)
 # export eth0 address for other scripts. This is assuming eth1 is the interface which is reachable from host network
 ip=$(/sbin/ip -o -4 addr list $1 | awk '{print $4}' | cut -d/ -f1)
-export ip=ip
 
 sudo apt-get update
 
@@ -40,21 +45,32 @@ sudo apt-get install -y unzip
 # zookeeper
 sudo apt-get install -y zookeeperd
 
-# install scala 2.11.7
-wget http://www.scala-lang.org/files/archive/scala-2.11.7.tgz
-sudo mkdir /usr/local/src/scala
-sudo tar xvf scala-2.11.7.tgz -C /usr/local/src/scala/
-rm scala-2.11.7.tgz
+
+if [ ! -d /usr/local/scala ]; then
+
+	# install scala 2.11.7
+	wget http://www.scala-lang.org/files/archive/scala-2.11.7.tgz
+	sudo mkdir /usr/local/src/scala
+	sudo tar xvf scala-2.11.7.tgz -C /usr/local/src/scala/
+
+	# remove scala tar file
+	rm scala-2.11.7.tgz
+fi
 
 # add to PATH env variable
 echo "export PATH=/usr/local/src/scala/scala-2.11.7/bin:$PATH" >> /etc/profile
 
 source /etc/profile
 
-# install spark 1.6.1 standalone
-wget https://d3kbcqa49mib13.cloudfront.net/spark-1.6.1-bin-hadoop2.6.tgz
-tar xzf spark-1.6.1-bin-hadoop2.6.tgz -C /usr/local/
-rm spark-1.6.1-bin-hadoop2.6.tgz
+if [ ! -d /usr/local/spark-1.6.1-bin-hadoop2.6 ]; then
+
+	# install spark 1.6.1 standalone
+	wget https://d3kbcqa49mib13.cloudfront.net/spark-1.6.1-bin-hadoop2.6.tgz
+	tar xzf spark-1.6.1-bin-hadoop2.6.tgz -C /usr/local/
+
+	# remove spark tar file
+	rm spark-1.6.1-bin-hadoop2.6.tgz
+fi
 
 # add to PATH env variable
 echo "export PATH=/usr/local/spark-1.6.1-bin-hadoop2.6/bin:$PATH" >> /etc/profile
@@ -63,43 +79,51 @@ source /etc/profile
 
 sudo pip install virtualenv
 
+# if proxy env variable is set, change the npm config
+if [[ ! (-z "${http_proxy}") || ! (-z "${HTTP_PROXY}") ]]; then
+	npm config set strict-ssl false
+	npm config set registry http://registry.npmjs.org/
+else
+	echo "No proxy required for npm"
+fi
+
 # install console-backend-utils
-bash ./console-backend.sh $pwd
+bash $pwd/scripts/console-backend.sh $pwd
 
 # install console-frontend
-bash ./console-frontend.sh $pwd
+bash $pwd/scripts/console-frontend.sh $pwd
 
 # install jupyter
-bash ./jupyter.sh $pwd
+bash $pwd/scripts/jupyter.sh $pwd
 
 # install hbase
-bash ./hbase.sh $pwd
+bash $pwd/scripts/hbase.sh $pwd
 
 # install kafka
-bash ./kafka.sh $pwd
+bash $pwd/scripts/kafka.sh $pwd
 
 # install jmxproxy
-bash ./jmxproxy.sh $pwd
+bash $pwd/scripts/jmxproxy.sh $pwd
 
 # install platform testing general
-bash ./platform-testing.sh $pwd
+bash $pwd/scripts/platform-testing.sh $pwd
 
 # install opentsdb
-bash ./opentsdb.sh $pwd
+bash $pwd/scripts/opentsdb.sh $pwd
 
 # install grafana
-bash ./grafana-server.sh $pwd
+bash $pwd/scripts/grafana-server.sh $pwd
 
 # install kafka-manager
-bash ./kafka-manager.sh $pwd
+bash $pwd/scripts/kafka-manager.sh $pwd
 
 # start opentsdb after a delay of 15 seconds and start kafka-manager
-cp $pwd/files/opentsdb-kafka-manager-boot.sh /opt/pnda
+cp $pwd/scripts/files/opentsdb-kafka-manager-boot.sh /opt/pnda
 
 # install crontab
 crontab -l > mycron
 #echo new cron into cron file
-echo "@reboot sleep 15 && /usr/local/hbase-1.2.0/bin/start-hbase.sh" >> mycron
+echo "@reboot sleep 10 && /usr/local/hbase-1.2.0/bin/start-hbase.sh" >> mycron
 echo "@reboot bash /opt/pnda/opentsdb-kafka-manager-boot.sh" >> mycron
 # start spark master & slave worker on reboot
 host_name=$(hostname)
@@ -109,17 +133,23 @@ echo "@reboot /usr/local/spark-1.6.1-bin-hadoop2.6/sbin/start-master.sh && /usr/
 crontab mycron
 rm mycron
 
-bash ./kafka-consumer.sh $pwd
+bash $pwd/scripts/kafka-consumer.sh $pwd
 
 # install useful libs
 sudo pip install pandas
 sudo pip install matplotlib
 sudo pip install scikit-learn
 
-bash ./platform-libraries.sh
+bash $pwd/scripts/platform-libraries.sh
 
-cp $pwd/files/zk-opentsdb-restart.sh /opt/pnda
+cp $pwd/scripts/files/zk-opentsdb-restart.sh /opt/pnda
 
-cp $pwd/assign-ip.sh /opt/pnda
+cp $pwd/scripts/assign-ip.sh /opt/pnda
 
 sudo bash /opt/pnda/assign-ip.sh $1
+
+echo "#####################################################"
+echo
+echo
+echo "Your Red-PNDA is successfully installed. Go to http://$ip on your browser to view the console!"
+
