+++ scripts/kafka.sh
@@ -1,15 +1,24 @@
 #!/bin/bash
 
-# kafka 0.10.2.0
-wget http://mirrors.gigenet.com/apache/kafka/0.10.2.0/kafka_2.11-0.10.2.0.tgz
-tar xzf kafka_2.11-0.10.2.0.tgz -C /usr/local/
-rm kafka_2.11-0.10.2.0.tgz
-cd /usr/local/kafka_2.11-0.10.2.0
+KAFKA_HOME=/usr/local/kafka_2.11-0.10.2.0
 
-# kafka will use  the standalone zookeeper instead of creating it's own
-# start kafka server
+# if directory does not exist
+if [ ! -d "$KAFKA_HOME" ]; then
+	# kafka 0.10.2.0
+	wget http://mirrors.gigenet.com/apache/kafka/0.10.2.0/kafka_2.11-0.10.2.0.tgz
+	tar xzf kafka_2.11-0.10.2.0.tgz -C /usr/local/
 
-cp $1/files/kafka.conf /etc/init/
+	# remove kafka tar
+	rm kafka_2.11-0.10.2.0.tgz
 
-# start kafka
-sudo service kafka start
+	cd /usr/local/kafka_2.11-0.10.2.0
+
+	# start kafka server
+
+	cp $1/scripts/files/kafka.conf /etc/init/
+
+	# start kafka
+	sudo service kafka start
+else
+	echo "Kafka already installed. Moving on!"
+fi
